fix: 超大目录分批移动性能问题

+++ internal/dbaccess/metadata_closure.go
@@ -59,13 +59,14 @@ func (m *metadataClosure) MoveSubTree(ctx context.Context, objectId, parentId st
 	// 2）查询 a 的所有后代：SELECT descendant FROM metadata_closure WHERE ancestor='%s'
 	// 3）组合2种情况：删除 a 和所有祖先的关系，以及 a 子树和所有祖先的关系。
 
-	sql := `select id,ancestor,descendant,depth from metadata_closure 
+	sql := fmt.Sprintf(`select id,ancestor,descendant,depth from metadata_closure 
 	where descendant IN (SELECT descendant FROM 
 						(SELECT descendant FROM metadata_closure WHERE ancestor='%s') as d)	--  后代节点(包括自己)
 		AND ancestor IN (SELECT ancestor FROM 
-						(SELECT ancestor FROM metadata_closure WHERE descendant='%s' AND ancestor != descendant) as a)	-- 祖先节点，不包括自己`
+						(SELECT ancestor FROM metadata_closure WHERE descendant='%s' AND ancestor != descendant) as a)	-- 祖先节点，不包括自己`,
+		objectId, objectId)
 
-	rows, err = tx.QueryContext(ctx, sql, objectId, objectId)
+	rows, err = tx.QueryContext(ctx, sql)
 	if err != nil {
 		return
 	}
@@ -82,20 +83,28 @@ func (m *metadataClosure) MoveSubTree(ctx context.Context, objectId, parentId st
 	}
 
 	// 删除和祖先的关系
-	args := make([]interface{}, 0, len(deletedClosure))
-	sqlPlacehoder := ""
-	for _, item := range deletedClosure {
-		sqlPlacehoder = "?,"
-		args = append(args, item.Id)
-	}
-	sql = fmt.Sprintf(`delete from metadata_closure where id in(%s)`, strings.Trim(sqlPlacehoder, ","))
-	row, err = tx.ExecContext(ctx, sql, args...)
-	if err != nil {
-		return
-	}
-	deleteCount, err = row.RowsAffected()
-	if err != nil {
-		return
+	if len(deletedClosure) > 0 {
+		args := make([]interface{}, 0, len(deletedClosure))
+		sqlPlacehoder := ""
+		const maxDelete = 10000
+		i := 0
+		for _, item := range deletedClosure {
+			sqlPlacehoder += "?,"
+			args = append(args, item.Id)
+			i++
+
+			if (i%maxDelete == 0) || (i+1 == len(deletedClosure)) {
+				sql = fmt.Sprintf(`delete from metadata_closure where id in(%s)`, strings.Trim(sqlPlacehoder, ","))
+				if row, err = tx.ExecContext(ctx, sql, args...); err != nil {
+					return
+				}
+				if deleteCount, err = row.RowsAffected(); err != nil {
+					return
+				}
+				args = make([]interface{}, 0, len(deletedClosure))
+				sqlPlacehoder = ""
+			}
+		}
 	}
 
 	// 插入到新的路径
@@ -152,10 +161,9 @@ func (m *metadataClosure) MoveSubTree(ctx context.Context, objectId, parentId st
 
 func (m *metadataClosure) MoveLargeSubTree(ctx context.Context, objectId, parentId string, pageSize int) (deleteCount, insertCount int64, err error) {
 	var (
-		rows  *sql.Rows  = nil
-		row   sql.Result = nil
-		tx    *sql.Tx    = nil
-		count int64      = 0
+		rows   *sql.Rows  = nil
+		result sql.Result = nil
+		count  int64      = 0
 	)
 
 	// 第一步：先断开 x 这个子树和祖先们的关系，x 变成孤立的树
@@ -182,54 +190,103 @@ func (m *metadataClosure) MoveLargeSubTree(ctx context.Context, objectId, parent
 	//
 	// 考虑到超大目录移动超时的情况，我们需要循环从叶子节点自底向上删除，将其变成孤立的树，插入时自上而下插入
 	//
-	for {
-		sqlStr := fmt.Sprintf(`select id from metadata_closure 
-			where descendant IN (SELECT descendant FROM 
-								(SELECT descendant FROM metadata_closure WHERE ancestor='%s') as d)	--  后代节点(包括自己)
-				AND ancestor IN (SELECT ancestor FROM 
-								(SELECT ancestor FROM metadata_closure WHERE descendant='%s' AND ancestor != descendant) as a)	-- 祖先节点，不包括自己
-				order by depth desc limit %d`, objectId, objectId, pageSize)
-
-		rows, err = m.db.QueryContext(ctx, sqlStr)
-		if err != nil {
-			return
-		}
-
-		// 删除和祖先的关系
-		args := make([]interface{}, 0, pageSize)
-		sqlPlacehoder := ""
-		for rows.Next() {
-			var id int = 0
-			if err = rows.Scan(&id); err != nil {
-				rows.Close()
-				return
-			}
-			sqlPlacehoder += "?,"
-			args = append(args, id)
-		}
-
-		if len(args) == 0 {
-			break
-		}
-
-		sqlStr = fmt.Sprintf(`delete from metadata_closure where id in(%s)`, strings.Trim(sqlPlacehoder, ","))
-		m.log.Debug("MoveLargeSubTree delete closure", zap.String("sql", sqlStr), zap.Any("args", args))
-
-		if row, err = m.db.ExecContext(ctx, sqlStr, args...); err != nil {
+	// 查询源的所有祖先，不包括自己
+	sqlStr := "select ancestor from metadata_closure where descendant=? and ancestor != descendant"
+	rows, err = m.db.QueryContext(ctx, sqlStr, objectId)
+	if err != nil {
+		return
+	}
+	srcParentPaths := make([]string, 0)
+	for rows.Next() {
+		var ancestor string
+		if err = rows.Scan(&ancestor); err != nil {
 			rows.Close()
 			return
 		}
-		if count, err = row.RowsAffected(); err != nil {
-			rows.Close()
-			return
-		}
+		srcParentPaths = append(srcParentPaths, ancestor)
+	}
+	rows.Close()
 
-		rows.Close()
-		deleteCount += count
+	// 查询后代，最大深度
+	sqlStr = "select max(depth) from metadata_closure where ancestor=?"
+	row := m.db.QueryRowContext(ctx, sqlStr, objectId)
+	maxDepth := 0
+	if err = row.Scan(&maxDepth); err != nil {
+		return
 	}
 
-	if deleteCount == 0 {
-		m.log.Warn("empty src subtree", zap.String("objectId", objectId), zap.String("parentId", parentId))
+	totalDescendantCount := 0
+
+	// 从叶子开始删除直到自己，如果此时没有查询到祖先，表明已经是一颗孤立的树
+	alreadyIsolatedTree := len(srcParentPaths) == 0
+
+	if !alreadyIsolatedTree {
+		// depth=0 自己到祖先也需要删除
+		for depth := maxDepth; depth >= 0; depth-- {
+			m.log.Info("MoveLargeSubTree delete depth", zap.Int("depth", depth), zap.Int("pageSize", pageSize))
+			lastDescendant := ""
+			for {
+				// 分页查询指定深度的后代
+				sqlStr := fmt.Sprintf(`select descendant from metadata_closure
+							where ancestor = '%s' and depth = %d and descendant > '%s'
+							  order by descendant asc limit %d`, objectId, depth, lastDescendant, pageSize)
+				if rows, err = m.db.QueryContext(ctx, sqlStr); err != nil {
+					return
+				}
+
+				// 读取分页查询的后代
+				descendants := make([]string, 0)
+				for rows.Next() {
+					var descendant string
+					if err = rows.Scan(&descendant); err != nil {
+						rows.Close()
+						return
+					}
+					descendants = append(descendants, descendant)
+					lastDescendant = descendant
+				}
+				rows.Close()
+
+				if len(descendants) == 0 {
+					break
+				} else {
+					totalDescendantCount += len(descendants)
+				}
+
+				// 和祖先的所有关系对
+				m.log.Info("MoveLargeSubTree delete closure start")
+				args1 := make([]string, len(srcParentPaths))
+				args2 := make([]string, len(descendants))
+				for i := range srcParentPaths {
+					args1[i] = fmt.Sprintf("'%s'", srcParentPaths[i])
+				}
+				for j := range descendants {
+					args2[j] = fmt.Sprintf("'%s'", descendants[j])
+				}
+
+				// 祖先相同的，一次性批量删除
+				sqlStr = fmt.Sprintf("delete from metadata_closure where ancestor in (%s) and descendant in (%s)",
+					strings.Join(args1, ","), strings.Join(args2, ","))
+				if result, err = m.db.ExecContext(ctx, sqlStr); err != nil {
+					return
+				}
+				if count, err = result.RowsAffected(); err != nil {
+					return
+				}
+				deleteCount += count
+
+				m.log.Info("MoveLargeSubTree delete closure end", zap.Int64("deleteCount", deleteCount))
+			}
+		}
+
+		if deleteCount == 0 {
+			m.log.Warn("empty src subtree", zap.String("objectId", objectId), zap.String("parentId", parentId))
+		} else {
+			m.log.Info(fmt.Sprintf("total delete closure: %d", deleteCount),
+				zap.Int("srcParentPathsCount", len(srcParentPaths)),
+				zap.Int("maxDescendantDepth", maxDepth),
+				zap.Int("totalDescendantCount", totalDescendantCount))
+		}
 	}
 
 	// 第二步：将上一步分离出的子树用笛卡尔积嫁接到新位置
@@ -239,70 +296,73 @@ func (m *metadataClosure) MoveLargeSubTree(ctx context.Context, objectId, parent
 	// 1) 查询目标父路径
 	// 2）查询孤立的树的后代，深度由小到大
 	// 3）取笛卡尔积，插入到新目标位置
-	sqlStr := "select ancestor, descendant, depth from metadata_closure where descendant=?"
+	sqlStr = "select ancestor, descendant, depth from metadata_closure where descendant=?"
 	rows, err = m.db.QueryContext(ctx, sqlStr, parentId)
 	if err != nil {
 		return
 	}
 
-	parentPaths := make([]interfaces.MetadataClosure, 0)
+	dstParentPaths := make([]interfaces.MetadataClosure, 0)
 	for rows.Next() {
 		item := interfaces.MetadataClosure{}
 		if err = rows.Scan(&item.Ancestor, &item.Descendant, &item.Depth); err != nil {
 			rows.Close()
 			return
 		}
-		parentPaths = append(parentPaths, item)
+		dstParentPaths = append(dstParentPaths, item)
 	}
 	rows.Close()
-	m.log.Info("MoveLargeSubTree before insert", zap.Any("parentPaths", parentPaths))
+	m.log.Info("MoveLargeSubTree before insert", zap.Any("dstParentPaths.depth", len(dstParentPaths)))
 
 	// 分批插入
-	startDescendant := ""
-	for {
-		// 查孤立树的所有后代，深度由小到大
-		sqlStr := `select descendant,depth from metadata_closure where 
-				   ancestor = ? and descendant > ? order by depth asc limit ?`
-		rows, err = m.db.QueryContext(ctx, sqlStr, objectId, startDescendant, pageSize)
-		if err != nil {
-			return
-		}
-
-		descendants := make([]interfaces.MetadataClosure, 0)
-		for rows.Next() {
-			item := interfaces.MetadataClosure{}
-			if err = rows.Scan(&item.Descendant, &item.Depth); err != nil {
-				rows.Close()
+	for depth := 0; depth <= maxDepth; depth++ {
+		startDescendant := ""
+		for {
+			// 查孤立树的所有后代，深度由小到大
+			sqlStr := `select descendant,depth from metadata_closure where 
+				   ancestor = ? and depth = ? and descendant > ? limit ?`
+			rows, err = m.db.QueryContext(ctx, sqlStr, objectId, depth, startDescendant, pageSize)
+			if err != nil {
 				return
 			}
-			descendants = append(descendants, item)
-			startDescendant = item.Descendant
-		}
-		rows.Close()
 
-		if len(descendants) == 0 {
-			break
-		}
-
-		// 取笛卡尔积插入
-		tx, err = m.db.Begin()
-		if err != nil {
-			return
-		}
-
-		for i := range parentPaths {
-			for j := range descendants {
-				sqlStr = fmt.Sprintf("insert into metadata_closure(ancestor,descendant,depth) values('%s','%s',%d)",
-					parentPaths[i].Ancestor, descendants[j].Descendant, parentPaths[i].Depth+descendants[j].Depth)
-
-				m.log.Debug("MoveLargeSubTree insert closure", zap.String("sql", sqlStr))
-
-				tx.ExecContext(ctx, sqlStr)
-				insertCount += 1
+			descendants := make([]interfaces.MetadataClosure, 0)
+			for rows.Next() {
+				item := interfaces.MetadataClosure{}
+				if err = rows.Scan(&item.Descendant, &item.Depth); err != nil {
+					rows.Close()
+					return
+				}
+				descendants = append(descendants, item)
+				startDescendant = item.Descendant
 			}
-		}
-		if err = tx.Commit(); err != nil {
-			return
+			rows.Close()
+
+			if len(descendants) == 0 {
+				break
+			}
+
+			// 取笛卡尔积插入
+			args := make([]string, 0)
+			for i := range dstParentPaths {
+				for j := range descendants {
+					args = append(args, fmt.Sprintf("('%s','%s',%d)",
+						dstParentPaths[i].Ancestor,
+						descendants[j].Descendant,
+						dstParentPaths[i].Depth+descendants[j].Depth+1))
+				}
+			}
+
+			// 一次性批量插入
+			result, err = m.db.ExecContext(ctx, "insert into metadata_closure(ancestor,descendant,depth) values"+strings.Join(args, ","))
+			if err != nil {
+				return
+			}
+			if count, err = result.RowsAffected(); err != nil {
+				return
+			}
+			insertCount += count
+			m.log.Debug("MoveLargeSubTree insert closure", zap.Int("count", len(dstParentPaths)*len(descendants)), zap.Int("depth", depth))
 		}
 	}
